pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "temporal-lora"
version = "0.1.0"
description = "Temporal LoRA for Dynamic Sentence Embeddings"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "torch>=2.2.0",
    "sentence-transformers>=3.0.0",
    "transformers>=4.35.0",
    "huggingface_hub>=0.20.0",
    "peft>=0.7.0",
    "datasets>=2.15.0",
    "faiss-cpu>=1.7.4",
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "umap-learn>=0.5.5",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "numpy>=1.26.0,<2.0.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "ruff==0.1.6",
    "black==23.11.0",
]

[project.scripts]
temporal-lora = "temporal_lora.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
# Prevent pytest-xdist which can cause FAISS issues on macOS
# Run tests serially to avoid threading conflicts
markers = [
    "faiss: tests that use FAISS (may be unstable on some systems)",
]
